apps/api/pyproject.toml
[project]
name = "archmentor-api"
version = "0.0.0"
description = "ArchMentor control-plane API (FastAPI)"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.32.1",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.7.0",
    "sqlmodel>=0.0.22",
    "alembic>=1.14.0",
    "psycopg[binary]>=3.2.3",
    "redis>=5.2.0",
    "httpx>=0.28.1",
    "pyjwt[crypto]>=2.10.1",
    "python-multipart>=0.0.17",
    "structlog>=24.4.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
    "ruff>=0.8.4",
    "ty>=0.0.1a11",
]

[build-system]
requires = ["uv_build>=0.5.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""
module-name = "archmentor_api"

[tool.ruff]
extend = "../../pyproject.toml"

[tool.ty.src]
root = "./archmentor_api"
